Eigen lib was migrated to latest: 24/01/2021

+++ src/tools.cpp
@@ -27,7 +27,7 @@ VectorXd Tools::CalculateRMSE(const vector<VectorXd> &estimations,
     }
 
     // accumulate squared residuals
-    for (int i = 0; i < estimations.size(); ++i)
+    for (size_t i = 0; i < estimations.size(); ++i)
     {
         VectorXd residual = estimations[i] - ground_truth[i];
         residual = residual.array() * residual.array();
